Move non protocol components back to top level
