+++ packages/wallet/src/components/network-status.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faDotCircle } from '@fortawesome/free-solid-svg-icons';
 import { connect } from 'react-redux';
-import { getNetworkId } from '../../../utils/contract-utils';
+import { getNetworkId } from '../utils/contract-utils';
 
 export class NetworkStatus extends React.PureComponent {
   render() {
