+++ packages/wallet/src/components/status-bar-layout.tsx
